Split Twitter scripts to always allow panel-based tipping
